6330 | 14 files | M DumpSource2/schemas/client/CDOTA_Modifier_Hoodwink_Sharpshooter_Windup.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Shredder_WhirlingDeath_Debuff.h,  M DumpSource2/schemas/client/CDOTA_Modifier_TemplarAssassin_Meld.h,  M DumpSource2/schemas/client/C_DOTA_Ability_Morphling_AdaptiveStrike_Agi.h,  M DumpSource2/schemas/server/CDOTA_Ability_Morphling_AdaptiveStrike_Agi.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Hoodwink_Sharpshooter_Windup.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Shredder_WhirlingDeath_Debuff.h,  M DumpSource2/schemas/server/CDOTA_Modifier_TemplarAssassin_Meld.h,  M DumpSource2/schemas/server/CDOTA_Tree.h,  M game/bin/built_from_cl.txt,  M game/dota/bin/linuxsteamrt64/libclient_strings.txt,  M game/dota/bin/linuxsteamrt64/libserver_strings.txt,  M game/dota/pak01_dir.txt,  M game/dota/steam.inf
https://steamdb.info/patchnotes/17463311/

+++ DumpSource2/schemas/server/CDOTA_Modifier_Shredder_WhirlingDeath_Debuff.h
@@ -1,7 +1,7 @@
 class CDOTA_Modifier_Shredder_WhirlingDeath_Debuff
 {
 	float32 stat_loss_pct;
-	int32 stat_loss_univ;
+	float32 stat_loss_univ;
 	int32 iPrimaryAttribute;
 	float32 flStatLossStr;
 	float32 flStatLossAgi;
